a minor refactor
