+++ src/components/GeneSetInformationColumn.js
@@ -5,8 +5,7 @@ import {GeneSetSubCohortBox} from './GeneSetSubCohortBox'
 import BaseStyle from '../css/base.css'
 import HoverGeneView from './hover/HoverGeneView'
 import SelectGeneView from './hover/SelectGeneView'
-
-const XENA_SS_LINK = 'https://xenabrowser.net/heatmap/'
+import {generateXenaLink} from '../functions/XenaLinkFunctions'
 
 export class GeneSetInformationColumn extends PureComponent {
 
@@ -28,90 +27,13 @@ export class GeneSetInformationColumn extends PureComponent {
 
   // ?columns=%5B%7B%22name%22%3A%22tcga_Kallisto_tpm%22%2C%22host%22%3A%22https%3A%2F%2Ftoil.xenahubs.net%22%2C%22fields%22%3A%22TP53%20FOXM1%22%7D%5D\">Example 1</a>"
 
-  generateLink(){
-    // const geneExpressionHost = getHostForGeneData(this.props.cohort[this.props.cohortIndex], this.props.view)
-    // const cnvHost = getHostForGeneData(this.props.cohort[this.props.cohortIndex], this.props.view)
-    // const mutationHost = getHostForGeneData(this.props.cohort[this.props.cohortIndex], this.props.view)
-
-    const cohort = this.props.cohort[this.props.cohortIndex]
-    // console.log('cohort',cohort)
-    // console.log('props',this.props)
-    const samples = this.props.pathwayData[this.props.cohortIndex].samples
-
-    const samplesJson = {
-      showWelcome: false,
-      searchSampleList: samples,
-    }
-    // filter: samples,
-
-    // console.log('samples',samples)
-    let geneExpressionDataset = cohort.geneExpression.dataset
-    let geneExpressionHost = cohort.geneExpression.host
-    let cnvDataset = cohort.copyNumberDataSetId
-    let cnvHost = cohort.host
-    const mutationDataSet= cohort.mutationDataSetId
-    const mutationHost = cohort.host
-
-    let linkString = ''
-
-    // console.log('gene expression ',geneExpressionDataset,geneExpressionHost)
-    // console.log('cnv dataset',cnvDataset,cnvHost)
-    // console.log('mutation dataset',mutationDataSet,mutationHost)
-    // add gene link
-    // let geneSetName = 'tcga_Kallisto_tpm'
-    // let geneSetHost = 'https://toil.xenahubs.net'
-    // let genes = 'TP53 FOXM1'
-    let genes = this.props.geneDataStats[this.props.cohortIndex].pathways.map( p => p.gene[0]).join(' ')
-    // console.log('genes',genes)
-    linkString += '?columns=['
-    linkString += `{"name":"${geneExpressionDataset}","host":"${geneExpressionHost}","fields":"${genes}","columnLabel":"Gene Set View","fieldLabel":"${genes.length} genes"}`
-    linkString += ','
-    linkString += `{"name":"${cnvDataset}","host":"${cnvHost}","fields":"${genes}","columnLabel":"CNV","fieldLabel":"${genes.length} genes"}`
-    linkString += ','
-    linkString += `{"name":"${mutationDataSet}","host":"${mutationHost}","fields":"${genes}","columnLabel":"Somatic Mutation","fieldLabel":"${genes.length} genes"}`
-
-    // show them individually
-    let count = 0
-    const geneSplit = genes.split(' ')
-    // console.log('# of genes ',geneSplit.length)
-    for(const g of geneSplit){
-      if(count < 1){
-        // console.log('gene',g)
-        if(count < geneSplit.length-1){
-          linkString += ','
-        }
-        linkString += `{"name":"${geneExpressionDataset}","host":"${geneExpressionHost}","fields":"${g}","columnLabel":"Gene ${g}","fieldLabel":"Individual Gene expression for ${g}"}`
-        linkString += ','
-        linkString += `{"name":"${cnvDataset}","host":"${cnvHost}","fields":"${g}","columnLabel":"Gene ${g}","fieldLabel":"Individual CNV ${g}"}`
-        linkString += ','
-        linkString += `{"name":"${mutationDataSet}","host":"${mutationHost}","fields":"${g}","columnLabel":"Gene ${g}","fieldLabel":"Individual Mutation ${g}"}`
-      }
-      ++count
-    }
-    linkString += ']'
-
-    // add gene link
-    // console.log('link string: ',linkString)
-
-    // <a id="link1" href="https://xenabrowser.net/heatmap/?columns=%5B%7B%22name%22%3A%22tcga_Kallisto_tpm%22%2C%22host%22%3A%22https%3A%2F%2Ftoil.xenahubs.net%22%2C%22fields%22%3A%22TP53%20FOXM1%22%7D%5D">Example 1</a>
-    // https://xenabrowser.net/heatmap/?columns=%5B%7B%22name%22:%22Gene%20Details%22,%22host%22:%22https://toil.xenahubs.net%22,%22fields%22:%22TP53%20FOXM1%22%7D%5D
-    // https://xenabrowser.net/heatmap/?columns=%5B%7B%22name%22%3A%22tcga_Kallisto_tpm%22%2C%22host%22%3A%22https%3A%2F%2Ftoil.xenahubs.net%22%2C%22fields%22%3A%22TP53%20FOXM1%22%7D%5D
-
-    let encodedUri = encodeURI(linkString)
-    // console.log('input samples json',samplesJson)
-    let heatmapUrl = '&heatmap='+encodeURI(JSON.stringify(samplesJson))
-    let finalLink = XENA_SS_LINK + encodedUri.replace(/:/g,'%3A').replace(/\//g,'%2F').replace(/,/g,'%2C')
-    finalLink += heatmapUrl
-    // console.log(finalLink)
-    return finalLink
-  }
 
   render() {
 
     const cohortColor = this.props.cohortColor[this.props.cohortIndex]
 
     if (this.props.geneDataStats && this.props.geneDataStats[this.props.cohortIndex].samples) {
-      const externalLink = this.generateLink()
+      const externalLink = generateXenaLink(this.props)
       return (
         <div
           className={BaseStyle.geneSetDetailBox}
@@ -126,7 +48,7 @@ export class GeneSetInformationColumn extends PureComponent {
               href={externalLink}
               rel="noopener noreferrer"
               target='_blank'
-              title={externalLink}>Link to SS</a>
+              title={externalLink}>Open in Xena</a>
           </div>
           <GeneSetSubCohortBox
             cohortIndex={this.props.cohortIndex}
